+++ CHANGES.rst
@@ -4,7 +4,12 @@ Changelog
 3.0.1 (unreleased)
 ------------------
 
-- bugfix: do not convert URIs such as 
+- Use ``application/javascript`` instead of ``application/x-javascript`` for
+  the ``Content-Type`` http header of JavaScript resources. Also see:
+  http://stackoverflow.com/questions/9664282
+  [thet]
+
+- bugfix: do not convert URIs such as
   url(data:image/svg+xml;base64,PD94bWwgdmVyc2lv) to absolute urls
   see https://github.com/plone/Products.ResourceRegistries/pull/16
   [cillianderoiste]